# api/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Qeet Pay API
  description: >-
    Qeet Pay — India-first payments, billing and GST infrastructure (UPI / Cards / NACH). Programmatic `/v1` calls
    authenticate with a `qp_live_` / `qp_test_` API key via the `X-Api-Key` header; money is always in integer minor
    units (paise).
  contact:
    name: Qeet Group
    url: https://qeet.in
  license:
    name: Proprietary
    url: https://qeet.in
  version: v1
servers:
  - url: https://api.pay.qeet.in
    description: Production
  - url: https://api.pay.staging.qeet.in
    description: Staging
  - url: http://localhost:4201
    description: Local (docker compose)
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/virtual-accounts:
    get:
      tags:
        - Virtual Accounts
      operationId: list_1
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountView'
    post:
      tags:
        - Virtual Accounts
      operationId: mint
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MintRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountView'
  /v1/virtual-accounts/{vaId}/credits:
    post:
      tags:
        - Virtual Accounts
      operationId: credit
      parameters:
        - name: vaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CreditView'
  /v1/virtual-accounts/{vaId}/close:
    post:
      tags:
        - Virtual Accounts
      operationId: close
      parameters:
        - name: vaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountView'
...
  /v1/settlements:
    get:
      tags:
        - Reconciliation
      operationId: list_3
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SettlementView'
    post:
      tags:
        - Reconciliation
      operationId: ingest
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestSettlementRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /v1/itc/reconcile:
    post:
      tags:
        - Input Tax Credit
      operationId: reconcile
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReconcileRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ReconSummary'
...
  /v1/virtual-accounts/{vaId}:
    get:
      tags:
        - Virtual Accounts
      operationId: get_1
      parameters:
        - name: vaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AccountWithCreditsView'
...
  /v1/settlements/{id}:
    get:
      tags:
        - Reconciliation
      operationId: get_3
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SettlementView'
...
components:
  schemas:
...
    MintRequest:
      type: object
      properties:
        customerRef:
          type: string
          minLength: 1
      required:
        - customerRef
    AccountView:
      type: object
      properties:
        id:
          type: string
        customerRef:
          type: string
        vaNumber:
          type: string
        ifsc:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
    CreditRequest:
      type: object
      properties:
        amountMinor:
          type: integer
          format: int64
        currency:
          type: string
          minLength: 1
        utr:
          type: string
          minLength: 1
        payerName:
          type: string
        payerRef:
          type: string
      required:
        - amountMinor
        - currency
        - utr
    CreditView:
      type: object
      properties:
        id:
          type: string
        vaId:
          type: string
        amountMinor:
          type: integer
          format: int64
        currency:
          type: string
        utr:
          type: string
        payerName:
          type: string
        payerRef:
          type: string
        ledgerEntryId:
          type: string
        creditedAt:
          type: string
          format: date-time
...
    IngestSettlementRequest:
      type: object
      properties:
        provider:
          type: string
          minLength: 1
        providerSettlementId:
          type: string
          minLength: 1
        currency:
          type: string
          minLength: 1
        settledAt:
          type: string
          format: date-time
        reportedNetMinor:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/LineDto'
          minItems: 1
      required:
        - currency
        - items
        - provider
        - providerSettlementId
    LineDto:
      type: object
      properties:
        paymentId:
          type: string
          format: uuid
        providerPaymentId:
          type: string
        grossMinor:
          type: integer
          format: int64
        feeMinor:
          type: integer
          format: int64
        taxMinor:
          type: integer
          format: int64
...
    Gstr2bLine:
      type: object
      properties:
        supplierGstin:
          type: string
        invoiceNumber:
          type: string
        totalGstMinor:
          type: integer
          format: int64
    ReconcileRequest:
      type: object
      properties:
        lines:
          type: array
          items:
            $ref: '#/components/schemas/Gstr2bLine'
      required:
        - lines
    ReconSummary:
      type: object
      properties:
        matched:
          type: integer
          format: int32
        mismatched:
          type: integer
          format: int32
        missingIn2b:
          type: integer
          format: int32
...
    AccountWithCreditsView:
      type: object
      properties:
        account:
          $ref: '#/components/schemas/AccountView'
        credits:
          type: array
          items:
            $ref: '#/components/schemas/CreditView'
...
    DiscrepancyView:
      type: object
      properties:
        type:
          type: string
        paymentId:
          type: string
        providerPaymentId:
          type: string
        expectedMinor:
          type: integer
          format: int64
        reportedMinor:
          type: integer
          format: int64
        detail:
          type: string
    ReconciliationView:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        matchedCount:
          type: integer
          format: int32
        discrepancyCount:
          type: integer
          format: int32
        discrepancies:
          type: array
          items:
            $ref: '#/components/schemas/DiscrepancyView'
    SettlementView:
      type: object
      properties:
        id:
          type: string
        provider:
          type: string
        providerSettlementId:
          type: string
        currency:
          type: string
        grossAmountMinor:
          type: integer
          format: int64
        feeAmountMinor:
          type: integer
          format: int64
        taxAmountMinor:
          type: integer
          format: int64
        netAmountMinor:
          type: integer
          format: int64
        itemCount:
          type: integer
          format: int32
        status:
          type: string
        ledgerEntryId:
          type: string
        reconciliation:
          $ref: '#/components/schemas/ReconciliationView'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      description: Programmatic API key (`qp_live_…` / `qp_test_…`) presented as the `X-Api-Key` header.
      name: X-Api-Key
      in: header
...